# Optimize get pending Safes query